+++ src/spatialdata/__init__.py
@@ -18,6 +18,7 @@ __all__ = [
     "concatenate",
     "rasterize",
     "to_circles",
+    "to_polygons",
     "transform",
     "aggregate",
     "bounding_box_query",
@@ -45,7 +46,7 @@ from spatialdata._core.data_extent import are_extents_equal, get_extent
 from spatialdata._core.operations.aggregate import aggregate
 from spatialdata._core.operations.rasterize import rasterize
 from spatialdata._core.operations.transform import transform
-from spatialdata._core.operations.vectorize import to_circles
+from spatialdata._core.operations.vectorize import to_circles, to_polygons
 from spatialdata._core.query._utils import circles_to_polygons, get_bounding_box_corners
 from spatialdata._core.query.relational_query import (
     get_values,